api/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
backcall==0.2.0
black==21.9b0
click==8.0.1; python_version >= '3.6'
decorator==5.1.0; python_version >= '3.5'
django-cors-headers==3.8.0
django-environ==0.7.0
django-extensions==3.1.3
django==3.2.7
djangorestframework==3.12.4
geojson==2.5.0
gunicorn==20.1.0
ipython==7.27.0
jedi==0.18.0; python_version >= '3.6'
markdown==3.3.4
matplotlib-inline==0.1.3; python_version >= '3.5'
mypy-extensions==0.4.3
parso==0.8.2; python_version >= '3.6'
pathspec==0.9.0
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
platformdirs==2.3.0; python_version >= '3.6'
prompt-toolkit==3.0.20; python_full_version >= '3.6.2'
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pygments==2.10.0; python_version >= '3.5'
pytz==2021.1
regex==2021.8.28
sqlparse==0.4.2; python_version >= '3.5'
tomli==1.2.1; python_version >= '3.6'
traitlets==5.1.0; python_version >= '3.7'
typing-extensions==3.10.0.2
wcwidth==0.2.5
